requirement.txt
altair==4.2.0
astor==0.8.1
attrs==19.3.0
backcall==0.1.0
backports.entry-points-selectable==1.1.1
backports.zoneinfo==0.2.1
base58==2.1.1
bleach==3.1.4
blinker==1.4
boto3==1.22.1
botocore==1.25.1
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==7.1.2
colorama==0.4.3
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
distlib==0.3.3
docopt==0.6.2
entrypoints==0.3
enum-compat==0.0.3
et-xmlfile==1.0.1
filelock==3.4.0
geographiclib==1.52
geopy==2.2.0
idna==3.3
ipykernel==5.2.1
ipython==7.13.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
jdcal==1.4.1
jedi==0.17.0
Jinja2==2.11.2
jmespath==1.0.0
joblib==0.17.0
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.3
jupyter-console==6.1.0
jupyter-core==4.6.3
kiwisolver==1.2.0
MarkupSafe==1.1.1
matplotlib==3.2.1
mistune==0.8.4
nbconvert==5.6.1
nbformat==5.0.5
nltk==3.5
notebook==6.0.3
numpy==1.18.2
openpyxl==3.0.3
packaging==21.3
pandas==1.0.3
pandocfilters==1.4.2
parso==0.7.0
pickleshare==0.7.5
Pillow==9.1.0
pipreqs==0.4.11
platformdirs==2.4.0
plotly==5.7.0
prometheus-client==0.7.1
prompt-toolkit==3.0.5
protobuf==3.20.1
pydeck==0.7.1
pyenchant==3.2.2
Pygments==2.6.1
PyInputPlus==0.2.9
pyparsing==2.4.7
pyrsistent==0.16.0
PySimpleValidate==0.2.10
python-dateutil==2.8.1
pytz==2019.3
pytz-deprecation-shim==0.1.0.post0
pywin32==227
pywinpty==0.5.7
pyzmq==19.0.0
qtconsole==4.7.3
QtPy==1.9.0
regex==2020.11.13
requests==2.27.1
s3transfer==0.5.2
scipy==1.8.0
seaborn==0.11.2
Send2Trash==1.5.0
six==1.14.0
stdiomask==0.0.5
streamlit==0.62.0
tenacity==8.0.1
terminado==0.8.3
testpath==0.4.4
toml==0.10.2
toolz==0.11.2
tornado==6.0.4
tqdm==4.54.0
traitlets==4.3.3
tzdata==2022.1
tzlocal==4.2
urllib3==1.26.9
validators==0.18.2
virtualenv==20.10.0
virtualenvwrapper-win==1.2.6
watchdog==2.1.7
wcwidth==0.1.9
webencodings==0.5.1
widgetsnbextension==3.5.1
xlrd==1.2.0
yarg==0.1.9
